requirements.txt
absl-py==2.2.2
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.2.4
asttokens==3.0.0
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
black==24.8.0
bleach==6.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
detectron2
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flake8==7.1.2
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.0
fvcore==0.1.5.post20221221
google-auth==2.40.1
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.1.0
iopath==0.1.9
ipykernel==6.29.5
ipython==8.12.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.7
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.7
llvmlite==0.31.0
lyft-dataset-sdk==0.0.8
markdown==3.7
markupsafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
mmcv-full==1.4.0
mmdet==2.14.0
# 最好自己安装
mmdetection3d 
mmsegmentation==0.14.1
mypy-extensions==1.1.0
narwhals==1.38.2
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
notebook==7.3.3
notebook-shim==0.2.4
numba==0.48.0
numpy==1.19.5
nuscenes-devkit==1.1.10
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==22.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pygments==2.19.1
pylint==3.2.7
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywavelets==1.4.1
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9.1
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
send2trash==1.8.3
setuptools==59.5.0
shapely==1.8.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.13
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.2
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==2.35.39
types-python-dateutil==2.9.0.20241206
typing-extensions==4.5.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.14
yacs==0.1.8
yapf==0.43.0
zipp==3.20.2
